fix: move portal to apps dir
