+++ apps/portal/src/app/app.config.ts
