+++ apps/portal/src/favicon.ico
