+++ apps/portal/tsconfig.json
@@ -22,7 +22,7 @@
       "path": "./tsconfig.editor.json"
     }
   ],
-  "extends": "../../../tsconfig.base.json",
+  "extends": "../../tsconfig.base.json",
   "angularCompilerOptions": {
     "enableI18nLegacyMessageIdFormat": false,
     "strictInjectionParameters": true,
