--- packages/portal/app/src/app/app.component.ts
@@ -1,11 +1,0 @@
-import { ChangeDetectionStrategy, Component, ViewEncapsulation } from '@angular/core';
-import { RouterOutlet } from '@angular/router';
-
-@Component({
-    imports: [RouterOutlet],
-    selector: 'app-root',
-    template: `<router-outlet />`,
-    encapsulation: ViewEncapsulation.None,
-    changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class AppComponent {}
